Replace "man" with "git help" and set COLUMNS
The "man git-some-sub-command" worked fine under Linux, but would result
in "No manual entry for..." on macOS. The "git help" works all around.

The man pages in the preview would "bleed" off the right side of the
page. Setting the COLUMN value to match the 75% that fzf is assigning to
it, fixes this issue.